# gen/ts-fetch/models/ControlPolicy.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ControlPolicy = {
    enabled?: boolean;
    cutoffPolicyId?: string;
    throttlingPolicyId?: string;
    cutoffThresholdKb?: number;
    cutoffThresholdMb?: number;
};
