# gen/ts-fetch/models/PackageListItem.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { PackageVersion } from './PackageVersion';
export type PackageListItem = {
    packageId?: string;
    name?: string;
    status?: string;
    latestVersion?: PackageVersion;
};
