# packages/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: IoT Connectivity Management Platform API
  description: >
    Core API for managing IoT SIM lifecycles, billing, and real-time connectivity diagnostics.
    Designed for MNOs, MVNOs, and Enterprise customers.
  version: 1.0.0
  contact:
    name: API Support
    email: support@iot-cmp.com
servers:
  - url: https://api.iot-cmp.com/v1
    description: Production Server
  - url: https://sandbox-api.iot-cmp.com/v1
    description: Sandbox Server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

    AdminApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Admin key for internal endpoints (same header name, different value).
  schemas:
...
    CommercialTerms:
      type: object
      properties:
        testPeriodDays:
          type: integer
        testQuotaKb:
          type: integer
        testExpiryCondition:
          type: string
          enum: [PERIOD_ONLY, QUOTA_ONLY, PERIOD_OR_QUOTA]
        testExpiryAction:
          type: string
          enum: [ACTIVATED, DEACTIVATED]
        commitmentPeriodMonths:
          type: integer
        commitmentPeriodDays:
          type: integer
    ControlPolicy:
      type: object
      properties:
        enabled:
          type: boolean
        cutoffPolicyId:
          type: string
        throttlingPolicyId:
          type: string
        cutoffThresholdKb:
          type: integer
        cutoffThresholdMb:
          type: integer
...
    RoamingProfileConfig:
      type: object
      properties:
        allowedMccMnc:
          type: array
          items:
            type: string
...
    CarrierServiceConfig:
      type: object
      required: [roamingProfileVersionId]
      properties:
        roamingProfileId:
          type: string
        supplierId:
          type: string
        carrierId:
          type: string
        rat:
          type: string
          enum: [4G, 3G, 5G, NB-IoT]
        apn:
          type: string
        apnProfileVersionId:
          type: string
        roamingProfileVersionId:
          type: string
        roamingProfile:
          $ref: '#/components/schemas/RoamingProfileConfig'
...
    PackageVersion:
      type: object
      properties:
        packageVersionId:
          type: string
        version:
          type: integer
        status:
          type: string
        effectiveFrom:
          type: string
          format: date-time
        supplierId:
          type: string
        operatorId:
          type: string
        carrierId:
          type: string
        serviceType:
          type: string
        apn:
          type: string
        roamingProfile:
          $ref: '#/components/schemas/RoamingProfileConfig'
        carrierServiceConfig:
          $ref: '#/components/schemas/CarrierServiceConfig'
        carrierServiceId:
          type: string
        controlPolicyId:
          type: string
        commercialTermsId:
          type: string
        controlPolicy:
          $ref: '#/components/schemas/ControlPolicy'
        commercialTerms:
          $ref: '#/components/schemas/CommercialTerms'
        pricePlanVersionId:
          type: string
        createdAt:
          type: string
          format: date-time
    PackageListItem:
      type: object
      properties:
        packageId:
          type: string
        enterpriseId:
          type: string
        name:
          type: string
        status:
          type: string
        latestVersion:
          $ref: '#/components/schemas/PackageVersion'
        createdAt:
          type: string
          format: date-time
    PackageListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PackageListItem'
        total:
          type: integer
...
paths:
...
  /packages:
    get:
      tags: [Packages]
      operationId: listPackagesByModuleRefs
      summary: Reverse Lookup Packages
      description: Reverse lookup packages by module references for tenant-visible scope.
      parameters:
        - name: pricePlanId
          in: query
          schema:
            type: string
          description: Reverse lookup by price plan id.
        - name: commercialTermsId
          in: query
          schema:
            type: string
          description: Reverse lookup by commercial terms id.
        - name: controlPolicyId
          in: query
          schema:
            type: string
          description: Reverse lookup by control policy id.
        - name: enterpriseId
          in: query
          schema:
            type: string
          description: Enterprise scope filter, required for reseller role.
        - name: status
          in: query
          schema:
            type: string
          description: Filter by package status.
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Reverse lookup package list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageListResponse'
...